Add CI workflows to fast-forward release branches upon tag push (#441)
* Add CI workflows to fast-forward release branches upon tag push

* Speed up the OS build step to free up disk space

* Update script to record installer versioning info in CI to match updates to install.planktoscope.community script

* Don't use deprecated paths in installer versioning info recording step

+++ software/distro/setup/ci-record-version.sh
@@ -145,9 +145,11 @@ main() {
     info "Recording versioning information to ${versioning_dir}..."
     if [ -d "${versioning_dir}" ]; then
       warn "The ${versioning_dir} directory already exists, so it will be erased."
-      sudo rm -rf "${versioning_dir}"
+      if ! rm -rf "${versioning_dir}" 2>/dev/null; then
+        sudo rm -rf "${versioning_dir}"
+      fi
     fi
-    if ! mkdir -p "${versioning_dir}"; then
+    if ! mkdir -p "${versioning_dir}" 2>/dev/null; then
       sudo mkdir -p "${versioning_dir}"
     fi
